refactor: changing Screens directory to Pages
My guess is Screens are more appropriate for small devices like mobile phones. Screen is likely
better to be used with react-native
