+++ src/App.js
@@ -2,8 +2,8 @@ import React from "react";
 // import * as BooksAPI from './BooksAPI'
 import "./assets/styles/App.css";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import Home from "./screens/Home";
-import Search from "./screens/Search";
+import Home from "./Pages/Home";
+import Search from "./Pages/Search";
 
 class BooksApp extends React.Component {
   render() {
